- implemented 'Play' button in Sandbox mode which allows you to play the crafted game state

+++ src/net/pferdimanzug/hearthstone/analyzer/utils/MathUtils.java
@@ -2,7 +2,7 @@ package net.pferdimanzug.hearthstone.analyzer.utils;
 
 public class MathUtils {
 
-	public static int clamp(int value, int min, int max) {
+	public static double clamp(double value, double min, double max) {
 		if (value < min) {
 			return min;
 		} else if (value > max) {
@@ -11,7 +11,7 @@ public class MathUtils {
 		return value;
 	}
 
-	public static double clamp(double value, double min, double max) {
+	public static int clamp(int value, int min, int max) {
 		if (value < min) {
 			return min;
 		} else if (value > max) {
